# docs/openapi.yaml
openapi: 3.0.3
info:
  title: NIX Platform API
  description: |
    API REST do NIX Platform — um monólito modular que centraliza
    integrações, automação e notificações. Todo endpoint autenticado espera
    um access token OIDC emitido pelo Keycloak como Bearer JWT; o backend o
    valida localmente contra um JWKS em cache (nunca com uma chamada por
    requisição ao Keycloak — ver §29 do documento de arquitetura).

    Toda resposta usa o mesmo envelope:
    `{"data": <payload|null>, "error": {"code","message"}|null, "meta": {...}|omitido}`.
    Erros nunca incluem stack trace ou outro detalhe interno.

    **Idempotência**: qualquer endpoint de escrita (POST/PUT/PATCH) aceita
    opcionalmente o header `Idempotency-Key` (ver
    `#/components/parameters/IdempotencyKey`). Reenviar a mesma requisição
    (mesmo método, caminho e corpo) com a mesma chave nunca executa o
    handler duas vezes — a segunda chamada em diante recebe de volta a
    resposta já produzida pela primeira (`Idempotent-Replay: true` no
    header da resposta), em vez de criar um segundo job/efeito colateral.
    Reusar a mesma chave com um corpo diferente é rejeitado com 409
    (`IDEMPOTENCY_KEY_REUSED`); uma segunda chamada enquanto a primeira
    ainda está em andamento recebe 409 (`IDEMPOTENCY_KEY_IN_PROGRESS`). Ver
    ADR [002](adr/002-enterprise-resilience-and-governance.md).

    **Resiliência e disponibilidade condicional**: chamadas que dependem
...
  version: "1.0.0"
  license:
    name: MIT
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/integrations/diario-oficial/test:
    post:
      tags: [diario-oficial]
      summary: Dispara uma verificação assíncrona de conectividade com o Diário Oficial
      operationId: testDiarioOficial
      description: >
        Não executa a verificação em si. Cria um job, o persiste junto com
        um evento de outbox em uma única transação, e retorna
        imediatamente — o worker o processa, e o frontend é notificado via
        WebSocket quando ele concluir ou falhar (§34/§72). Sujeito a rate
        limiting por chamador. Aceita `Idempotency-Key` — recomendado para
        evitar criar jobs duplicados num duplo clique ou retry automático
        do cliente.
      security:
        - bearerAuth: [integrations:test]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "202":
          $ref: "#/components/responses/TestJobAccepted"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/IdempotencyConflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/IntegrationUnavailable"
  /api/v1/diario-oficial/monitored-terms:
    post:
      tags: [diario-oficial]
      summary: Cadastra um termo pra monitorar no Diário Oficial (MVP real via DJEN)
      operationId: createMonitoredTerm
      description: >
        Pelo menos um critério de busca é obrigatório: `oab_number`+`oab_uf` (sempre juntos),
        `process_number`, ou `free_text`. O worker sincroniza todo termo ATIVO contra o DJEN
        (Diário de Justiça Eletrônico Nacional, CNJ) a cada 6h — este endpoint só cadastra, nunca
        busca no DJEN diretamente.
      security:
        - bearerAuth: [diario_oficial:manage]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                label: { type: string, example: "Dr. Fulano — OAB/MG 419" }
                oab_number: { type: string, example: "419" }
                oab_uf: { type: string, example: "MG" }
                process_number: { type: string, example: "50015349420258130351" }
                free_text: { type: string }
              required: [label]
      responses:
        "201":
          description: Termo cadastrado.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/MonitoredTerm" }
        "400":
          description: Corpo inválido (JSON malformado).
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          description: >
            Nenhum critério de busca preenchido, ou `oab_number` sem `oab_uf` (ou vice-versa).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
    get:
      tags: [diario-oficial]
      summary: Lista todo termo monitorado, mais recente primeiro
      operationId: listMonitoredTerms
      security:
        - bearerAuth: [diario_oficial:read]
      responses:
        "200":
          description: Lista de termos monitorados.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/MonitoredTerm" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/diario-oficial/monitored-terms/{termID}:
    delete:
      tags: [diario-oficial]
      summary: Remove um termo monitorado
      operationId: deleteMonitoredTerm
      description: >
        Publicações/matches já encontrados pra este termo são removidos em cascata — o histórico de
        um termo apagado não continua aparecendo no feed de publicações recentes. Responde 200 com
        um envelope vazio (nunca 204 sem corpo — o proxy BFF/cliente desta plataforma sempre espera
        um Envelope JSON na resposta).
      security:
        - bearerAuth: [diario_oficial:manage]
      parameters:
        - name: termID
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Termo removido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "400":
          description: termID não é um UUID válido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Termo não encontrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /api/v1/diario-oficial/monitored-terms/{termID}/publications:
    get:
      tags: [diario-oficial]
      summary: Lista as publicações que casaram com um termo, mais recentes primeiro
      operationId: listPublicationsForTerm
      security:
        - bearerAuth: [diario_oficial:read]
      parameters:
        - name: termID
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: >
            Página de publicações casadas — um termID inexistente devolve uma página vazia, não 404
            (a mesma escolha que o resto desta API já faz pra listagens escopadas por id).
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/MatchedPublication" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/diario-oficial/publications:
    get:
      tags: [diario-oficial]
      summary: Feed agregado de publicações casadas entre TODO termo monitorado
      operationId: listRecentPublications
      security:
        - bearerAuth: [diario_oficial:read]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Página de publicações casadas, mais recentes primeiro.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/MatchedPublication" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/v1/scanning/projects/{projectID}/findings/{fingerprint}/triage:
    put:
      tags: [scanning]
      summary: Tria um achado deduplicado (Fase 14 — Maturidade de AppSec)
      operationId: triageFinding
      description: >
        Marca o fingerprint como false_positive/wont_fix/risk_accepted dentro do projeto — chamar de
        novo no MESMO fingerprint substitui a decisão anterior, não acumula histórico. reason é
        obrigatório: uma supressão sem justificativa registrada é o tipo de decisão que uma auditoria
        de segurança depois cobra explicação. Gravado em audit_logs (finding.triaged).
      security:
        - bearerAuth: [scanning:manage]
      parameters:
        - name: projectID
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: fingerprint
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status, reason]
              properties:
                status:
                  type: string
                  enum: [false_positive, wont_fix, risk_accepted]
                reason:
                  type: string
                  description: Obrigatório — motivo da triagem, no máximo 2000 caracteres.
                expires_at:
                  type: string
                  format: date-time
                  description: >
                    Opcional (Fase 14, continuação — expiração de triagem). Precisa estar no futuro.
                    Quando o prazo passa, o achado volta a contar como aberto em
                    SecurityPosture/ListProjectFindingsHistory (mas a decisão em si nunca é apagada
                    automaticamente — ver GET .../findings-history's triage_expired).
      responses:
        "200":
          description: Triagem gravada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "400":
          description: projectID/fingerprint inválido, status desconhecido, reason vazio/longo demais, ou expires_at no passado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Projeto não encontrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
    delete:
      tags: [scanning]
      summary: Reabre um achado já triado (Fase 14 — Maturidade de AppSec)
      operationId: untriageFinding
      description: >
        Apaga a decisão de triagem — idempotente, chamar num fingerprint nunca triado (ou já
        reaberto) não é erro. Gravado em audit_logs (finding.untriaged).
      security:
        - bearerAuth: [scanning:manage]
      parameters:
        - name: projectID
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: fingerprint
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Achado reaberto (ou já estava aberto).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "400":
          description: projectID/fingerprint inválido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Um access token OIDC emitido pelo Keycloak. Validado localmente
        (assinatura, emissor, audiência, expiração, algoritmo) contra um
        JWKS em cache.
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: >
        Chave única por operação lógica (recomenda-se um UUID gerado no
        cliente), escopada ao usuário autenticado. Reenviar a mesma
        requisição com a mesma chave reproduz a resposta original em vez
        de reexecutar a operação — ver a nota de Idempotência na descrição
        geral da API.
      schema:
        type: string
      example: "3f2b9e0a-6d34-4e77-9c3e-2b6f8e6b6a3d"
  responses:
    TestJobAccepted:
      description: >
        O teste foi enfileirado. Se a requisição carregou um
        `Idempotency-Key` já usado com sucesso antes (mesmo corpo), esta é
        uma resposta reproduzida (replay) da chamada original — o header
        `Idempotent-Replay: true` distingue esse caso, e nenhum job novo
        foi criado.
      headers:
        Idempotent-Replay:
          description: Presente e "true" somente quando esta resposta é um replay de uma chamada anterior com a mesma Idempotency-Key.
          schema:
            type: string
            enum: ["true"]
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/Envelope"
              - type: object
                properties:
                  data:
                    type: object
                    properties:
                      job_id:
                        type: string
                        format: uuid
                      status:
                        type: string
                        enum: [queued]
          example:
            data: { job_id: "3c1f1e2a-...", status: "queued" }
            error: null
    IdempotencyConflict:
      description: >
        Conflito relacionado à Idempotency-Key: `IDEMPOTENCY_KEY_IN_PROGRESS`
        se uma requisição com a mesma chave ainda está sendo processada, ou
        `IDEMPOTENCY_KEY_REUSED` se a chave já foi usada com um corpo de
        requisição diferente.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
          examples:
            in_progress:
              summary: Requisição original ainda em andamento
              value:
                data: null
                error: { code: "IDEMPOTENCY_KEY_IN_PROGRESS", message: "a request with this Idempotency-Key is already being processed" }
            reused:
              summary: Mesma chave usada com um corpo diferente
              value:
                data: null
                error: { code: "IDEMPOTENCY_KEY_REUSED", message: "Idempotency-Key was already used with a different request" }
    IntegrationUnavailable:
      description: >
        A integração externa está temporariamente indisponível.
        `CIRCUIT_OPEN` (o circuit breaker interno abriu depois de falhas
        consecutivas — tende a se resolver sozinho) e `FEATURE_DISABLED`
        (um nix-admin desligou a integração via feature flag — só volta
        quando religada) usam o mesmo status 503, diferenciados pelo
        `error.code`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
          examples:
            circuit_open:
              summary: Circuit breaker aberto
              value:
                data: null
                error: { code: "CIRCUIT_OPEN", message: "diario-oficial: temporarily unavailable (circuit breaker open)" }
            feature_disabled:
              summary: Feature flag desabilitada por um administrador
              value:
                data: null
                error: { code: "FEATURE_DISABLED", message: "the \"diario_oficial_scraping_enabled\" feature is currently disabled" }
    Unauthorized:
      description: Access token ausente, inválido ou expirado.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
          example:
            data: null
            error: { code: "UNAUTHORIZED", message: "invalid or expired access token" }
    Forbidden:
      description: Autenticado, mas sem o role/permissão necessário.
      content:
        application/json:
          example:
            data: null
            error: { code: "FORBIDDEN", message: "insufficient permission" }
...
    TooManyRequests:
      description: Limite de requisições excedido.
      content:
        application/json:
          example:
            data: null
            error: { code: "RATE_LIMITED", message: "too many requests, please try again later" }
  schemas:
    Envelope:
      type: object
      description: >
        O formato base de toda resposta. Operações bem-sucedidas restringem
        `data` a um tipo concreto via `allOf` no próprio schema da resposta;
        esta forma base documenta o contrato genérico {data, error, meta}.
      properties:
        data:
          type: object
          nullable: true
        error:
          type: object
          nullable: true
          allOf:
            - $ref: "#/components/schemas/ErrorBody"
        meta:
          type: object
          nullable: true
      required: [data, error]
    ErrorBody:
      type: object
      properties:
        code:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "email: failed on \"required\""
      required: [code, message]
...
    MonitoredTerm:
      type: object
      description: >
        MVP de monitoramento real do Diário Oficial via DJEN — o que o usuário quer acompanhar.
        Pelo menos um critério é preenchido: oab_number+oab_uf, process_number, ou free_text.
      properties:
        id: { type: string, format: uuid }
        label: { type: string, example: "Dr. Fulano — OAB/MG 419" }
        oab_number: { type: string }
        oab_uf: { type: string }
        process_number: { type: string }
        free_text: { type: string }
        active: { type: boolean }
        last_synced_at:
          type: string
          format: date-time
          description: Ausente se este termo ainda nunca foi sincronizado pelo worker.
        created_at: { type: string, format: date-time }
      required: [id, label, active, created_at]
    MatchedPublication:
      type: object
      description: >
        Uma publicação do DJEN que casou com um MonitoredTerm — persistida pelo worker
        (nunca buscada sob demanda), com o resultado do casamento junto.
      properties:
        id: { type: string, format: uuid }
        tribunal: { type: string, example: "TJMG" }
        orgao: { type: string }
        tipo_comunicacao: { type: string, example: "Intimação" }
        texto: { type: string }
        process_number: { type: string }
        process_number_masked: { type: string, example: "5001534-94.2025.8.13.0351" }
        availability_date: { type: string, format: date }
        link: { type: string, description: "URL da publicação original no sistema do tribunal." }
        monitored_term_id: { type: string, format: uuid }
        monitored_term_label: { type: string }
        matched_at: { type: string, format: date-time }
      required: [id, tribunal, orgao, tipo_comunicacao, texto, process_number, process_number_masked, availability_date, monitored_term_id, monitored_term_label, matched_at]
...